backend/requirements.txt
Django==4.2
djangorestframework==3.15
pandas==2.2.2
reportlab==4.0
django-cors-headers==3.14.0
